requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_text_splitters
lxml
langchain_openai
chromadb
langchain-chroma
faiss-cpu
sentence_transformers
langchain_huggingface
langchain_groq
langserve
Flask
Flask-SQLAlchemy
flask-cors
flask-migrate
Flask-Mail